tests/test-requirements.txt
# Testing Requirements for YOUTILITY5 AI Features
# Core testing framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0  # For parallel test execution
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Test data and factories
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2  # For time-based testing

# Performance and load testing
pytest-benchmark==4.0.0
locust==2.17.0  # For load testing AI endpoints

# Database testing
pytest-postgresql==5.0.0
pytest-redis==3.0.2

# HTTP testing
responses==0.24.1  # For mocking HTTP requests
requests-mock==1.11.0

# AI/ML testing utilities
#unittest-mock==1.0.1  # Additional mocking utilities
scikit-learn==1.3.2  # For testing ML components (already in main requirements)
numpy>=1.26.0  # For numerical testing (Python 3.12 compatible)
pandas==2.1.4  # For data testing

# Celery testing
celery[redis]==5.3.4  # For async task testing
kombu==5.3.4

# Security testing
bandit==1.7.5  # For security vulnerability scanning
safety==2.3.5  # For dependency vulnerability scanning

# Coverage and reporting
coverage==7.3.2
pytest-html==4.1.1  # For HTML test reports
pytest-json-report==1.5.0  # For JSON test reports

# Static analysis for tests
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Memory profiling for performance tests
memory-profiler==0.61.0
psutil==5.9.6

# Image processing testing (for Quality Assurance tests)
Pillow==10.1.0  # For image quality testing
opencv-python-headless==4.8.1.78  # For computer vision testing

# Text processing testing (for NLP tests)
textblob==0.17.1
nltk==3.8.1

# Time zone testing
pytz==2023.3

# Parallel execution
pytest-parallel==0.1.1

# GraphQL testing
graphene-django-test==0.1.1
snapshottest==0.6.0  # For GraphQL snapshot testing

# API testing utilities  
httpx==0.25.2  # For async HTTP testing
pytest-httpx==0.28.0
djangorestframework-testing==0.1.0

# WebSocket testing (for real-time features)
channels-testing==0.2.0

# Rate limiting testing
django-ratelimit==4.1.0
